backend/requirements.txt
# FastAPI web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database driver and connection pooling
pymysql==1.1.0
DBUtils==3.0.3

# Configuration management
pydantic-settings==2.1.0

# OAuth & JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0

# AI - Embedding Model
sentence-transformers==2.2.2

# AI - Vector Database
chromadb==0.4.18

# AI - LLM (for local inference)
torch==2.1.1
transformers==4.35.2
accelerate==0.25.0
bitsandbytes==0.41.3

# Utilities
python-dotenv==1.0.0